exit

+++ estimate.py
@@ -66,7 +66,7 @@ def monte_carlo(n):
         if ((x * 2) + (y * 2)) <= 1:
             inside += 1
         i += 1
-    pi = float((4 * inside) / n)
+    pi = (4 * inside) / n
     return pi
